+++ M2/Macaulay2/e/aring-zz-flint.hpp
@@ -34,9 +34,9 @@ namespace M2 {
     
     size_t cardinality() const { return static_cast<size_t>(-1); }
 
-    unsigned long computeHashValue(const ElementType& a) const 
+    unsigned int computeHashValue(const ElementType& a) const 
     { 
-      return fmpz_get_ui(&a);
+      return static_cast<unsigned int>(fmpz_get_ui(&a));
     }
     
     /** @name properties
